refact: product reviews post release cleanup - rename v2 files to remove v2 suffix
